delete useless RegisterModel funtion
