+++ edge/pkg/edged/edged.go
@@ -104,7 +104,6 @@ import (
 	utilpod "github.com/kubeedge/kubeedge/edge/pkg/edged/util/pod"
 	"github.com/kubeedge/kubeedge/edge/pkg/edged/util/record"
 	csiplugin "github.com/kubeedge/kubeedge/edge/pkg/edged/volume/csi"
-	"github.com/kubeedge/kubeedge/edge/pkg/metamanager"
 	"github.com/kubeedge/kubeedge/edge/pkg/metamanager/client"
 	"github.com/kubeedge/kubeedge/pkg/version"
 )
@@ -832,7 +831,7 @@ func (e *edged) syncPod() {
 	//send msg to metamanager to get existing pods
 	info := model.NewMessage("").BuildRouter(e.Name(), e.Group(), e.namespace+"/"+model.ResourceTypePod,
 		model.QueryOperation)
-	beehiveContext.Send(metamanager.MetaManagerModuleName, *info)
+	beehiveContext.Send(constants.MetaManagerModuleName, *info)
 	for {
 		select {
 		case <-beehiveContext.Done():
@@ -868,7 +867,7 @@ func (e *edged) syncPod() {
 		klog.Infof("result content is %s", result.Content)
 		switch resType {
 		case model.ResourceTypePod:
-			if op == model.ResponseOperation && resID == "" && result.GetSource() == metamanager.MetaManagerModuleName {
+			if op == model.ResponseOperation && resID == "" && result.GetSource() == constants.MetaManagerModuleName {
 				err := e.handlePodListFromMetaManager(content)
 				if err != nil {
 					klog.Errorf("handle podList failed: %v", err)
