+++ edge/pkg/metamanager/client/persistentvolume.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/kubeedge/beehive/pkg/core/model"
 	"github.com/kubeedge/kubeedge/common/constants"
-	"github.com/kubeedge/kubeedge/edge/pkg/metamanager"
 )
 
 // PersistentVolumesGetter is interface to get client PersistentVolumes
@@ -68,7 +67,7 @@ func (c *persistentvolumes) Get(name string, options metav1.GetOptions) (*api.Pe
 		}
 	}
 
-	if msg.GetOperation() == model.ResponseOperation && msg.GetSource() == metamanager.MetaManagerModuleName {
+	if msg.GetOperation() == model.ResponseOperation && msg.GetSource() == constants.MetaManagerModuleName {
 		return handlePersistentVolumeFromMetaDB(content)
 	}
 	return handlePersistentVolumeFromMetaManager(content)
