+++ edge/pkg/metamanager/process.go
@@ -38,7 +38,7 @@ func feedbackError(err error, info string, request model.Message) {
 	if err != nil {
 		errInfo = fmt.Sprintf(info+": %v", err)
 	}
-	errResponse := model.NewErrorMessage(&request, errInfo).SetRoute(MetaManagerModuleName, request.GetGroup())
+	errResponse := model.NewErrorMessage(&request, errInfo).SetRoute(constants.MetaManagerModuleName, request.GetGroup())
 	if request.GetSource() == constants.EdgedModuleName {
 		sendToEdged(errResponse, request.IsSync())
 	} else {
@@ -365,7 +365,7 @@ func (m *metaManager) processQuery(message model.Message) {
 			m.processRemoteQuery(message)
 		} else {
 			resp := message.NewRespByMessage(&message, *metas)
-			resp.SetRoute(MetaManagerModuleName, resp.GetGroup())
+			resp.SetRoute(constants.MetaManagerModuleName, resp.GetGroup())
 			sendToEdged(resp, message.IsSync())
 		}
 		return
@@ -382,7 +382,7 @@ func (m *metaManager) processQuery(message model.Message) {
 		feedbackError(err, "Error to query meta in DB", message)
 	} else {
 		resp := message.NewRespByMessage(&message, *metas)
-		resp.SetRoute(MetaManagerModuleName, resp.GetGroup())
+		resp.SetRoute(constants.MetaManagerModuleName, resp.GetGroup())
 		if resType == constants.ResourceTypeService || resType == constants.ResourceTypeEndpoints || resType == model.ResourceTypePodlist {
 			sendToEdgeMesh(resp, message.IsSync())
 		} else {
@@ -493,7 +493,7 @@ func (m *metaManager) syncPodStatus() {
 		content = append(content, podStatus)
 	}
 
-	msg := model.NewMessage("").BuildRouter(MetaManagerModuleName, GroupResource, namespace+constants.ResourceSep+model.ResourceTypePodStatus, model.UpdateOperation).FillBody(content)
+	msg := model.NewMessage("").BuildRouter(constants.MetaManagerModuleName, GroupResource, namespace+constants.ResourceSep+model.ResourceTypePodStatus, model.UpdateOperation).FillBody(content)
 	sendToCloud(msg)
 	klog.Infof("sync pod status successful, %s", msgDebugInfo(msg))
 }
